LayoutTests/platform/chromium/accessibility/scroll-to-make-visible-with-subfocus-expected.txt

Tests that scrolling to make a certain region within an element visible successfully scrolls the main window.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


5000-pixel box
Target
PASS window.pageYOffset is 0
PASS window.pageYOffset is >= 6982
PASS 7582 is >= window.pageYOffset
PASS successfullyParsed is true

TEST COMPLETE
